+++ src/components/PopoverOption/index.js
@@ -1,0 +1,1 @@
+export { PopoverOption } from './PopoverOption'
